Restructure district queries
