+++ src/components/CreateOrgs.vue
@@ -209,7 +209,7 @@ import _without from 'lodash/without';
 import { useVuelidate } from '@vuelidate/core';
 import { required, requiredIf } from '@vuelidate/validators';
 import { useAuthStore } from '@/store/auth';
-import useDistrictsQuery from '@/composables/queries/useDistrictsQuery';
+import useDistrictsListQuery from '@/composables/queries/useDistrictsListQuery';
 import useDistrictSchoolsQuery from '@/composables/queries/useDistrictSchoolsQuery';
 import useSchoolClassesQuery from '@/composables/queries/useSchoolClassesQuery';
 import useGroupsQuery from '@/composables/queries/useGroupsQuery';
@@ -247,7 +247,7 @@ onMounted(() => {
   if (roarfirekit.value.restConfig) initTable();
 });
 
-const { isLoading: isLoadingDistricts, data: districts } = useDistrictsQuery({
+const { isLoading: isLoadingDistricts, data: districts } = useDistrictsListQuery({
   enabled: initialized,
 });
