--- src/composables/queries/useDistrictQuery.js
@@ -1,28 +1,0 @@
-import { useQuery } from '@tanstack/vue-query';
-import _isEmpty from 'lodash/isEmpty';
-import { computeQueryOverrides } from '@/helpers/computeQueryOverrides';
-import { fetchDocById } from '@/helpers/query/utils';
-import { DISTRICT_QUERY_KEY } from '@/constants/queryKeys';
-import { FIRESTORE_COLLECTIONS } from '@/constants/firebase';
-
-/**
- * District query.
- *
- * @param {String} districtId – The ID of the district to fetch.
- * @param {QueryOptions|undefined} queryOptions – Optional TanStack query options.
- * @returns {UseQueryResult} The TanStack query result.
- */
-const useDistrictQuery = (districtId, queryOptions = undefined) => {
-  // Ensure all necessary data is loaded before enabling the query.
-  const conditions = [() => !_isEmpty(districtId)];
-  const { isQueryEnabled, options } = computeQueryOverrides(conditions, queryOptions);
-
-  return useQuery({
-    queryKey: [DISTRICT_QUERY_KEY, districtId],
-    queryFn: () => fetchDocById(FIRESTORE_COLLECTIONS.DISTRICTS, districtId),
-    enabled: isQueryEnabled,
-    ...options,
-  });
-};
-
-export default useDistrictQuery;
