+++ src/composables/queries/useDistrictsListQuery.js
@@ -1,0 +1,40 @@
+import { computed } from 'vue';
+import { useQuery } from '@tanstack/vue-query';
+import _isEmpty from 'lodash/isEmpty';
+import { computeQueryOverrides } from '@/helpers/computeQueryOverrides';
+import { orgFetcher } from '@/helpers/query/orgs';
+import useUserClaimsQuery from '@/composables/queries/useUserClaimsQuery';
+import useUserType from '@/composables/useUserType';
+import { DISTRICTS_LIST_QUERY_KEY } from '@/constants/queryKeys';
+import { FIRESTORE_COLLECTIONS } from '@/constants/firebase';
+
+/**
+ * All Districts query.
+ *
+ * @param {QueryOptions|undefined} queryOptions – Optional TanStack query options.
+ * @returns {UseQueryResult} The TanStack query result.
+ */
+const useDistrictsListQuery = (queryOptions = undefined) => {
+  // Fetch the user claims.
+  const { data: userClaims } = useUserClaimsQuery({
+    enabled: queryOptions?.enabled ?? true,
+  });
+
+  // Get admin status and administation orgs.
+  const { isSuperAdmin } = useUserType(userClaims);
+  const administrationOrgs = computed(() => userClaims.value?.claims?.minimalAdminOrgs);
+
+  // Ensure all necessary data is loaded before enabling the query.
+  const claimsLoaded = computed(() => !_isEmpty(userClaims?.value?.claims));
+  const queryConditions = [() => claimsLoaded.value];
+  const { isQueryEnabled, options } = computeQueryOverrides(queryConditions, queryOptions);
+
+  return useQuery({
+    queryKey: [DISTRICTS_LIST_QUERY_KEY],
+    queryFn: () => orgFetcher(FIRESTORE_COLLECTIONS.DISTRICTS, undefined, isSuperAdmin, administrationOrgs),
+    enabled: isQueryEnabled,
+    ...options,
+  });
+};
+
+export default useDistrictsListQuery;
