+++ src/composables/queries/useDistrictsQuery.js
@@ -1,37 +1,26 @@
-import { computed } from 'vue';
+import { toValue } from 'vue';
 import { useQuery } from '@tanstack/vue-query';
 import _isEmpty from 'lodash/isEmpty';
 import { computeQueryOverrides } from '@/helpers/computeQueryOverrides';
-import { orgFetcher } from '@/helpers/query/orgs';
-import useUserClaimsQuery from '@/composables/queries/useUserClaimsQuery';
-import useUserType from '@/composables/useUserType';
+import { fetchDocumentsById } from '@/helpers/query/utils';
 import { DISTRICTS_QUERY_KEY } from '@/constants/queryKeys';
 import { FIRESTORE_COLLECTIONS } from '@/constants/firebase';
 
 /**
  * Districts query.
  *
+ * @param {Array} districtIds – The array of class IDs to fetch.
  * @param {QueryOptions|undefined} queryOptions – Optional TanStack query options.
  * @returns {UseQueryResult} The TanStack query result.
  */
-const useDistrictsQuery = (queryOptions = undefined) => {
-  // Fetch the user claims.
-  const { data: userClaims } = useUserClaimsQuery({
-    enabled: queryOptions?.enabled ?? true,
-  });
-
-  // Get admin status and administation orgs.
-  const { isSuperAdmin } = useUserType(userClaims);
-  const administrationOrgs = computed(() => userClaims.value?.claims?.minimalAdminOrgs);
-
+const useDistrictsQuery = (districtIds, queryOptions = undefined) => {
   // Ensure all necessary data is loaded before enabling the query.
-  const claimsLoaded = computed(() => !_isEmpty(userClaims?.value?.claims));
-  const queryConditions = [() => claimsLoaded.value];
-  const { isQueryEnabled, options } = computeQueryOverrides(queryConditions, queryOptions);
+  const conditions = [() => !_isEmpty(districtIds)];
+  const { isQueryEnabled, options } = computeQueryOverrides(conditions, queryOptions);
 
   return useQuery({
-    queryKey: [DISTRICTS_QUERY_KEY],
-    queryFn: () => orgFetcher(FIRESTORE_COLLECTIONS.DISTRICTS, undefined, isSuperAdmin, administrationOrgs),
+    queryKey: [DISTRICTS_QUERY_KEY, districtIds],
+    queryFn: () => fetchDocumentsById(FIRESTORE_COLLECTIONS.DISTRICTS, toValue(districtIds)),
     enabled: isQueryEnabled,
     ...options,
   });
