+++ src/composables/queries/useDistrictsQuery.test.js
@@ -1,16 +1,13 @@
-import { ref } from 'vue';
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import * as VueQuery from '@tanstack/vue-query';
+import { nanoid } from 'nanoid';
 import { withSetup } from '@/test-support/withSetup.js';
-import { orgFetcher } from '@/helpers/query/orgs';
+import { fetchDocumentsById } from '@/helpers/query/utils';
 import useDistrictsQuery from './useDistrictsQuery';
-import useUserClaimsQuery from '@/composables/queries/useUserClaimsQuery';
 
-vi.mock('@/helpers/query/orgs', () => ({
-  orgFetcher: vi.fn().mockImplementation(() => []),
+vi.mock('@/helpers/query/utils', () => ({
+  fetchDocumentsById: vi.fn().mockImplementation(() => []),
 }));
-
-vi.mock('@/composables/queries/useUserClaimsQuery');
 
 vi.mock('@tanstack/vue-query', async (getModule) => {
   const original = await getModule();
@@ -23,81 +20,71 @@ vi.mock('@tanstack/vue-query', async (getModule) => {
 describe('useDistrictsQuery', () => {
   let queryClient;
 
-  const mockUserClaims = ref({
-    claims: {
-      minimalAdminOrgs: ['mock-org-id-1', 'mock-org-id-2'],
-      super_admin: true,
-    },
-  });
-
   beforeEach(() => {
     queryClient = new VueQuery.QueryClient();
   });
 
   afterEach(() => {
     queryClient?.clear();
-    vi.clearAllMocks();
   });
 
-  it('should call query with correct parameters', () => {
-    vi.mocked(useUserClaimsQuery).mockReturnValue({ data: mockUserClaims });
+  it('should call query with correct parameters when fetching a specific district', () => {
+    const districtIds = nanoid();
+
     vi.spyOn(VueQuery, 'useQuery');
 
-    withSetup(() => useDistrictsQuery(), {
+    withSetup(() => useDistrictsQuery(districtIds), {
       plugins: [[VueQuery.VueQueryPlugin, { queryClient }]],
     });
 
     expect(VueQuery.useQuery).toHaveBeenCalledWith({
-      queryKey: ['districts'],
+      queryKey: ['districts', districtIds],
       queryFn: expect.any(Function),
       enabled: expect.objectContaining({
         _value: true,
       }),
     });
 
-    expect(orgFetcher).toHaveBeenCalledWith(
-      'districts',
-      undefined,
-      expect.objectContaining({ value: true }),
-      expect.objectContaining({ value: ['mock-org-id-1', 'mock-org-id-2'] }),
-    );
-  });
-
-  it('should only fetch districts only once user claims are loaded', async () => {
-    vi.mocked(useUserClaimsQuery).mockReturnValue({ data: {}, isLoading: ref(true) });
-
-    vi.spyOn(VueQuery, 'useQuery');
-
-    withSetup(() => useDistrictsQuery(), {
-      plugins: [[VueQuery.VueQueryPlugin, { queryClient }]],
-    });
-
-    expect(VueQuery.useQuery).toHaveBeenCalledWith({
-      queryKey: ['districts'],
-      queryFn: expect.any(Function),
-      enabled: expect.objectContaining({
-        _value: false,
-      }),
-    });
-
-    expect(orgFetcher).not.toHaveBeenCalled();
+    expect(fetchDocumentsById).toHaveBeenCalledWith('districts', districtIds);
   });
 
   it('should allow the query to be disabled via the passed query options', () => {
+    const districtIds = nanoid();
     const queryOptions = { enabled: false };
 
     vi.spyOn(VueQuery, 'useQuery');
 
-    withSetup(() => useDistrictsQuery(queryOptions), {
+    withSetup(() => useDistrictsQuery(districtIds, queryOptions), {
       plugins: [[VueQuery.VueQueryPlugin, { queryClient }]],
     });
 
     expect(VueQuery.useQuery).toHaveBeenCalledWith({
-      queryKey: ['districts'],
+      queryKey: ['districts', districtIds],
       queryFn: expect.any(Function),
       enabled: expect.objectContaining({
         _value: false,
       }),
     });
+  });
+
+  it('should keep the query disabled if not district IDs are specified', () => {
+    const districtIds = [];
+    const queryOptions = { enabled: true };
+
+    vi.spyOn(VueQuery, 'useQuery');
+
+    withSetup(() => useDistrictsQuery(districtIds, queryOptions), {
+      plugins: [[VueQuery.VueQueryPlugin, { queryClient }]],
+    });
+
+    expect(VueQuery.useQuery).toHaveBeenCalledWith({
+      queryKey: ['districts', districtIds],
+      queryFn: expect.any(Function),
+      enabled: expect.objectContaining({
+        _value: false,
+      }),
+    });
+
+    expect(fetchDocumentsById).not.toHaveBeenCalled();
   });
 });
