+++ src/constants/queryKeys.js
@@ -2,8 +2,8 @@ export const ADMINISTRATIONS_QUERY_KEY = 'administrations';
 export const ADMINISTRATIONS_LIST_QUERY_KEY = 'administrations-list';
 export const ADMINISTRATION_ASSIGNMENTS_QUERY_KEY = 'administration-assignments';
 export const CLASSES_QUERY_KEY = 'classes';
-export const DISTRICT_QUERY_KEY = 'district';
 export const DISTRICTS_QUERY_KEY = 'districts';
+export const DISTRICTS_LIST_QUERY_KEY = 'districts-list';
 export const DISTRICT_SCHOOLS_QUERY_KEY = 'district-schools';
 export const DSGF_ORGS_QUERY_KEY = 'dsgf-orgs';
 export const FAMILY_QUERY_KEY = 'family';
